+++ AptiWithAB-API/src/main/java/com/aptiwithabapi/daos/ResultDao.java
@@ -7,10 +7,10 @@ import com.aptiwithabapi.models.Result;
 public interface ResultDao {
 
 	List<Result>	getAllResults();
-	List<Result>	getAllResultsForTest(int testId);
+	List<Result>	getAllResultsForTest(long testId);
 	List<Result>	getAllResultsForStudent(int prn);
 	Result			create(Result result);
 	boolean			deleteAllResults();
-	boolean			deleteAllResultsForTest(int testId);
+	boolean			deleteAllResultsForTest(long testId);
 	boolean			deleteAllResultsForStudent(int prn);
 }
